Reduce duplication between demo and tests
The demo and the tests now use precisely the same infrastructure to boot up
nodes. This also significantly cleans up the tests.
