+++ ouroboros-consensus/src/Ouroboros/Consensus/Util/STM.hs
@@ -4,6 +4,8 @@
 module Ouroboros.Consensus.Util.STM (
     -- * Misc
     monitorTVar
+  , blockUntilJust
+  , blockUntilAllJust
     -- * Simulate various monad stacks in STM
   , Sim
   , simId
@@ -41,6 +43,16 @@ monitorTVar f initB tvar notify = fork $ go initB
                      else return (a, b')
       notify a
       go b'
+
+blockUntilJust :: MonadSTM m => TVar m (Maybe a) -> Tr m a
+blockUntilJust var = do
+    ma <- readTVar var
+    case ma of
+      Nothing -> retry
+      Just a  -> return a
+
+blockUntilAllJust :: MonadSTM m => [TVar m (Maybe a)] -> Tr m [a]
+blockUntilAllJust = mapM blockUntilJust
 
 {-------------------------------------------------------------------------------
   Simulate monad stacks
