+++ ouroboros-consensus/test-consensus/Test/Dynamic/BFT.hs
@@ -28,13 +28,8 @@ import           Test.QuickCheck
 import           Test.Tasty
 import           Test.Tasty.QuickCheck
 
-import           Ouroboros.Consensus.Crypto.DSIGN.Mock
-import           Ouroboros.Consensus.Ledger.Abstract
-import           Ouroboros.Consensus.Ledger.Mock
+import           Ouroboros.Consensus.Demo
 import           Ouroboros.Consensus.Node
-import           Ouroboros.Consensus.Protocol.Abstract
-import           Ouroboros.Consensus.Protocol.BFT
-import           Ouroboros.Consensus.Protocol.ExtNodeConfig
 import           Ouroboros.Consensus.Util.Random
 import           Ouroboros.Network.Chain (Chain)
 
@@ -45,44 +40,14 @@ tests = testGroup "Dynamic chain generation" [
       testProperty "simple BFT convergence" prop_simple_bft_convergence
     ]
 
-
-type Protocol = Bft BftMockCrypto
-
-instance HasPayload (Bft BftMockCrypto) (BlockUnderTest Protocol) where
-  blockPayload _ = testPayloadP
-                 . encPayloadP
-                 . headerOuroboros
-                 . simpleHeader
-
-instance ProtocolLedgerView (BlockUnderTest Protocol) where
-  protocolLedgerView (EncNodeConfig _ cfg) (SimpleLedgerState u _) =
-      ((), nodeStake cfg u)
-
 prop_simple_bft_convergence :: Seed -> Property
 prop_simple_bft_convergence =
-    prop_simple_protocol_convergence mkConfig mkState initialChainState isValid
+    prop_simple_protocol_convergence
+      (protocolInfo DemoBFT (NumCoreNodes numNodes))
+      isValid
   where
-    mkConfig :: Int -> NodeConfig Protocol
-    mkConfig i = BftNodeConfig
-        { bftNodeId   = CoreId i
-        , bftSignKey  = SignKeyMockDSIGN i
-        , bftNumNodes = fromIntegral numNodes
-        , bftVerKeys  = verKeys
-        }
-      where
-        verKeys :: Map NodeId (VerKeyDSIGN (BftDSIGN BftMockCrypto))
-        verKeys = Map.fromList [ (CoreId j, VerKeyMockDSIGN j)
-                               | j <- [0 .. numNodes - 1]
-                               ]
-
-    mkState :: Int -> NodeState Protocol
-    mkState _ = ()
-
-    initialChainState :: ChainState Protocol
-    initialChainState = ()
-
     isValid :: [NodeId]
-            -> [(VTime, Map NodeId (Chain (BlockUnderTest Protocol)))]
+            -> [(VTime, Map NodeId (Chain (Block DemoBFT)))]
             -> Property
     isValid nodeIds trace = counterexample (show trace) $
       case trace of
@@ -91,5 +56,5 @@ prop_simple_bft_convergence =
                      .&&. allEqual (takeChainPrefix <$> Map.elems final)
         _otherwise   -> property False
       where
-        takeChainPrefix :: Chain (BlockUnderTest Protocol) -> Chain (BlockUnderTest Protocol)
+        takeChainPrefix :: Chain (Block DemoBFT) -> Chain (Block DemoBFT)
         takeChainPrefix = id -- in BFT, chains should indeed all be equal.
